+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 ### SuffixForge compatibility:
 * Removed the SuffixForge compatibility module (`soulbound` will still be disabled if SuffixForge is present)
 
+### Citizens compatibility:
+* "Players" created with the Citizens plugin are no longer treated as a real player
+
 ## 1.5.0-BETA
 ### General changes:
 * Added a `soulbound` command which marks items as soulbound, preventing them from being dropped on death